+++ example-projects/plugin-config/src/Source.java
@@ -1,0 +1,5 @@
+// [impl->dsn~md-exampleA~1]
+// [impl->dsn~asciidoc-exampleB~1]
+class Source
+{
+}